Create main.dart
course solution generates a fixed set of three widgets for three coin types, so adding a new coin to cryptoList would involve adding another widget for new coin. my solution adapted code to generate list of coins with a for loop (instead of creating a separate element for each coin). so new coins only need to be added to list in the future and no other changes will be needed.